requirements.txt
# Holo V1 MVP Dependencies
# Python 3.11+

# API Framework
fastapi==0.115.*
uvicorn[standard]==0.34.*
pydantic==2.*
pydantic-settings==2.*

# LLM Providers (needed from Step 5 onward, install now)
openai>=1.0
anthropic>=0.40
google-genai>=1.0

# Database
supabase>=2.0

# Streamlit UI (needed from Step 10 onward, install now)
streamlit>=1.42
requests>=2.31

# Utilities
python-dotenv>=1.0

# Testing
pytest>=8.0
httpx>=0.27
